+++ app/components/Review/ReviewSectionField.tsx
@@ -20,13 +20,15 @@ const ReviewSectionField: React.FC<FieldProps> = (props) => {
     () => Object.keys(errorSchema || {}).length > 0 || !!hasFormContextError,
     [errorSchema, hasFormContextError]
   );
+  const toggledSection = formContext?.toggledSection === pageName;
   const allowAnalystEdit =
     uiOptions?.allowAnalystEdit && (formContext.isEditable ?? true);
   return (
     <Accordion
       id={idSchema.$id}
       name={name}
-      defaultToggled={uiOptions.defaultExpanded || hasErrors}
+      defaultToggled={uiOptions.defaultExpanded || hasErrors || toggledSection}
+      focused={toggledSection}
       toggled={formContext.toggleOverride}
       cbcId={formContext.cbcId}
       isCBC={formContext.isCBC}