+++ app/tests/pages/analyst/application/[applicationId]/rfi/index.test.tsx
@@ -265,6 +265,23 @@ describe('The index page', () => {
     ).toBeVisible();
   });
 
+  it('should show indications for template 1 and 2', async () => {
+    pageTestingHelper.loadQuery();
+    pageTestingHelper.renderPage();
+
+    expect(
+      screen.getByText(
+        /RFI upload for Template 1 automatically updates the data for Final Eligible Households and Indigenous/
+      )
+    ).toBeInTheDocument();
+
+    expect(
+      screen.getByText(
+        /FI upload for Template 2 automatically updates the data for Total Eligible Costs and Total Project Costs/
+      )
+    ).toBeInTheDocument();
+  });
+
   it('shows all of the correct Requested file items', async () => {
     pageTestingHelper.loadQuery();
     pageTestingHelper.renderPage();